refactor：拆分重构长文件，使用vite进行打包 enhance：优化通知
