+++ .gitignore
@@ -1,0 +1,52 @@
+# 依赖文件
+node_modules/
+npm-debug.log
+yarn-debug.log
+yarn-error.log
+package-lock.json
+yarn.lock
+
+# 构建输出
+dist/
+build/
+*.zip
+
+# 开发环境
+.env
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+# IDE 和编辑器
+.idea/
+.vscode/
+*.swp
+*.swo
+.DS_Store
+*.sublime*
+.history/
+
+# 缓存文件
+.cache/
+.parcel-cache/
+.npm/
+.eslintcache
+
+# 测试覆盖率
+coverage/
+
+# Chrome扩展开发生成的临时文件
+*.pem
+*.crx
+
+# Vite 构建相关
+.vite/
+
+# 日志文件
+logs/
+*.log
+
+# 其他临时文件
+temp/
+tmp/
