+++ offscreen.html
@@ -8,6 +8,6 @@
   <!-- 用于解析HTML和播放音频的离屏文档 -->
   <div id="parser-container" style="display:none;"></div>
   <!-- 音频播放器将在需要时由JavaScript创建 -->
-  <script src="js/offscreen.js"></script>
+  <script type="module" src="./js/offscreen.js"></script>
 </body>
 </html>
