+++ package.json
@@ -1,0 +1,28 @@
+{
+  "name": "leva",
+  "version": "1.0.0",
+  "description": "这是一个帮助用户收集和高亮显示网页上生词的浏览器扩展。该扩展允许用户通过右键菜单将选中的单词添加到生词本中，并能在网页上自动高亮显示这些已收集的生词，帮助用户巩固记忆。同时，提供单词悬浮释义功能，辅助用户理解和学习单词。",
+  "type": "module",
+  "main": "index.js",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/wothn/manifest.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/wothn/manifest/issues"
+  },
+  "homepage": "https://github.com/wothn/manifest#readme",
+  "devDependencies": {
+    "@crxjs/vite-plugin": "^1.0.14",
+    "vite": "^2.9.15"
+  }
+}
