+++ popup.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset="utf-8">
   <title>生词本</title>
-  <link rel="stylesheet" href="css/popup.css">
+  <link rel="stylesheet" href="./css/popup.css">
 </head>
 <body>
   <div class="container">
@@ -21,6 +21,6 @@
       <ul id="wordsList"></ul>
     </div>
   </div>
-  <script src="js/popup.js"></script>
+  <script type="module" src="./js/popup.js"></script>
 </body>
 </html>
